add client side redirect
