+++ apps/devportal/next.config.js
@@ -3,9 +3,6 @@
  * @type {import('next').NextConfig}
  */
 
-const fs = require('fs');
-const path = require('path');
-
 //const withTM = require('next-transpile-modules'); // pass the modules you would like to see transpiled
 const securityHeaders = [
   {
@@ -34,6 +31,7 @@ const securityHeaders = [
     value: '1; mode=block',
   },
 ];
+
 const redirects = [
   {
     source: '/learn/integrations/xm-cdp',
@@ -140,31 +138,7 @@ const nextConfig = {
     ];
   },
   async redirects() {
-    // Get the latest newsletter and redirect `newsletter/latest` to it
-    const newsletterDataDir = path.resolve(__dirname, 'data/newsletters/');
-    const year = fs
-      .readdirSync(newsletterDataDir)
-      .map((y) => parseInt(y, 10))
-      .sort((a, b) => b - a)[0];
-    const month = fs
-      .readdirSync(path.resolve(newsletterDataDir, `${year}`))
-      .map((m) => {
-        const name = m.substring(m, m.length - 5);
-        return {
-          name,
-          num: parseInt(name, 10),
-        };
-      })
-      .sort((a, b) => b.num - a.num)[0].name;
-
-    return [
-      ...redirects,
-      {
-        source: '/newsletter/latest',
-        destination: `/newsletter/${year}/${month}`,
-        permanent: false,
-      },
-    ];
+    return [...redirects];
   },
 };
